+++ aztk/node_scripts/install/install.py
@@ -1,10 +1,12 @@
 import os
-from core import config
-from install import pick_master, spark, scripts, create_user, plugins, spark_container
+
 import wait_until_master_selected
-from aztk.models.plugins import PluginTarget
 from aztk.internal import cluster_data
-from .node_scheduling import setup_node_scheduling
+from aztk.models.plugins import PluginTarget
+from core import config
+from install import create_user, pick_master, plugins, spark, spark_container
+
+# from .node_scheduling import setup_node_scheduling
 
 
 def read_cluster_config():
@@ -14,14 +16,16 @@ def read_cluster_config():
     return cluster_config
 
 
-def setup_host(docker_repo: str):
+def setup_host(docker_repo: str, docker_run_options: str):
     """
-    Code to be run on the node(NOT in a container)
+    Code to be run on the node (NOT in a container)
+    :param docker_repo: location of the Docker image to use
+    :param docker_run_options: additional command-line options to pass to docker run
     """
     client = config.batch_client
 
     create_user.create_user(batch_client=client)
-    if os.environ['AZ_BATCH_NODE_IS_DEDICATED'] == "true" or os.environ['AZTK_MIXED_MODE'] == "false":
+    if os.environ["AZ_BATCH_NODE_IS_DEDICATED"] == "true" or os.environ["AZTK_MIXED_MODE"] == "false":
         is_master = pick_master.find_master(client)
     else:
         is_master = False
@@ -44,11 +48,12 @@ def setup_host(docker_repo: str):
 
     cluster_conf = read_cluster_config()
 
-    setup_node_scheduling(client, cluster_conf, is_master)
+    # setup_node_scheduling(client, cluster_conf, is_master)
 
-    #TODO pass azure file shares
+    # TODO pass azure file shares
     spark_container.start_spark_container(
         docker_repo=docker_repo,
+        docker_run_options=docker_run_options,
         gpu_enabled=os.environ.get("AZTK_GPU_ENABLED") == "true",
         plugins=cluster_conf.plugins,
     )
@@ -76,6 +81,5 @@ def setup_spark_container():
         spark.start_spark_worker()
 
     plugins.setup_plugins(target=PluginTarget.SparkContainer, is_master=is_master, is_worker=is_worker)
-    scripts.run_custom_scripts(is_master=is_master, is_worker=is_worker)
 
-    open("/tmp/setup_complete", 'a').close()
+    open("/tmp/setup_complete", "a").close()