+++ aztk/spark/client/cluster/helpers/create.py
@@ -12,19 +12,18 @@ POOL_ADMIN_USER_IDENTITY = batch_models.UserIdentity(
     auto_user=batch_models.AutoUserSpecification(
         scope=batch_models.AutoUserScope.pool, elevation_level=batch_models.ElevationLevel.admin))
 
-
-def _default_scheduling_target(vm_count: int):
-    if vm_count == 0:
-        return models.SchedulingTarget.Any
-    else:
-        return models.SchedulingTarget.Dedicated
+# def _default_scheduling_target(vm_count: int):
+#     if vm_count == 0:
+#         return models.SchedulingTarget.Any
+#     else:
+#         return models.SchedulingTarget.Dedicated
 
 
 def _apply_default_for_cluster_config(configuration: models.ClusterConfiguration):
     cluster_conf = models.ClusterConfiguration()
     cluster_conf.merge(configuration)
-    if cluster_conf.scheduling_target is None:
-        cluster_conf.scheduling_target = _default_scheduling_target(cluster_conf.size)
+    # if cluster_conf.scheduling_target is None:
+    #     cluster_conf.scheduling_target = _default_scheduling_target(cluster_conf.size)
     return cluster_conf
 
 
@@ -52,9 +51,16 @@ def create_cluster(core_cluster_operations,
         zip_resource_files = cluster_data.upload_node_data(node_data).to_resource_file()
 
         start_task = spark_cluster_operations._generate_cluster_start_task(
-            core_cluster_operations, zip_resource_files, cluster_conf.cluster_id, cluster_conf.gpu_enabled(),
-            cluster_conf.get_docker_repo(), cluster_conf.file_shares, cluster_conf.plugins, cluster_conf.mixed_mode(),
-            cluster_conf.worker_on_master)
+            core_cluster_operations,
+            zip_resource_files,
+            cluster_conf.cluster_id,
+            cluster_conf.gpu_enabled(),
+            cluster_conf.get_docker_repo(),
+            cluster_conf.get_docker_run_options(),
+            cluster_conf.file_shares,
+            cluster_conf.mixed_mode(),
+            cluster_conf.worker_on_master,
+        )
 
         software_metadata_key = base_models.Software.spark