+++ aztk_cli/spark/endpoints/cluster/cluster_create.py
@@ -3,33 +3,32 @@ import typing
 
 import aztk.spark
 from aztk.spark.models import ClusterConfiguration, UserConfiguration
-from aztk.utils import deprecate
 from aztk_cli import config, log, utils
 from aztk_cli.config import load_aztk_spark_config
 
 
 def setup_parser(parser: argparse.ArgumentParser):
-    parser.add_argument('--id', dest='cluster_id', help='The unique id of your spark cluster')
-    parser.add_argument('--size', type=int, help='Number of vms in your cluster')
+    parser.add_argument("--id", dest="cluster_id", help="The unique id of your spark cluster")
+    parser.add_argument("--size", type=int, help="Number of vms in your cluster")
     parser.add_argument(
-        '--size-low-pri',
+        "--size-low-pri",
         type=int,
-        help='Number of low priority vms in your cluster (Deprecated, use --size-low-priority)')
-    parser.add_argument('--size-low-priority', type=int, help='Number of low priority vms in your cluster')
-    parser.add_argument('--vm-size', help='VM size for nodes in your cluster')
-    parser.add_argument('--username', help='Username to access your cluster (required: --wait flag)')
+        help="Number of low priority vms in your cluster (Deprecated, use --size-low-priority)",
+    )
+    parser.add_argument("--size-low-priority", type=int, help="Number of low priority vms in your cluster")
+    parser.add_argument("--vm-size", help="VM size for nodes in your cluster")
+    parser.add_argument("--username", help="Username to access your cluster (required: --wait flag)")
     parser.add_argument(
-        '--password',
-        help="The password to access your spark cluster's head \
-                             node. If not provided will use ssh public key.")
+        "--password",
+        help="The password to access your spark cluster's head node. If not provided will use ssh public key.",
+    )
     parser.add_argument(
-        '--docker-repo',
-        help='The location of the public docker image you want to use \
-                             (<my-username>/<my-repo>:<tag>)')
-    parser.add_argument('--subnet-id', help='The subnet in which to create the cluster.')
+        "--docker-repo", help="The location of the public docker image you want to use (<my-username>/<my-repo>:<tag>)")
+    parser.add_argument("--docker-run-options", help="command line options to pass to `docker run`")
+    parser.add_argument("--subnet-id", help="The subnet in which to create the cluster.")
 
-    parser.add_argument('--no-wait', dest='wait', action='store_false')
-    parser.add_argument('--wait', dest='wait', action='store_true')
+    parser.add_argument("--no-wait", dest="wait", action="store_false")
+    parser.add_argument("--wait", dest="wait", action="store_true")
     parser.set_defaults(wait=None, size=None, size_low_pri=None, size_low_priority=None)
 
 
@@ -41,9 +40,6 @@ def execute(args: typing.NamedTuple):
     # read cluster.yaml configuration file, overwrite values with args
     file_config, wait = config.read_cluster_config()
     cluster_conf.merge(file_config)
-    if args.size_low_pri is not None:
-        deprecate("0.9.0", "--size-low-pri has been deprecated.", "Please use --size-low-priority.")
-        args.size_low_priority = args.size_low_pri
 
     cluster_conf.merge(
         ClusterConfiguration(
@@ -52,22 +48,26 @@ def execute(args: typing.NamedTuple):
             size_low_priority=args.size_low_priority,
             vm_size=args.vm_size,
             subnet_id=args.subnet_id,
-            user_configuration=UserConfiguration(
-                username=args.username,
-                password=args.password,
-            )))
+            user_configuration=UserConfiguration(username=args.username, password=args.password),
+        ))
 
-    if args.docker_repo and cluster_conf.toolkit:
-        cluster_conf.toolkit.docker_repo = args.docker_repo
+    if cluster_conf.toolkit:
+        if args.docker_repo:
+            cluster_conf.toolkit.docker_repo = args.docker_repo
+        if args.docker_run_options:
+            cluster_conf.toolkit.docker_run_options = args.docker_run_options
 
     wait = wait if args.wait is None else args.wait
 
     user_configuration = cluster_conf.user_configuration
 
     if user_configuration and user_configuration.username:
-        ssh_key, password = utils.get_ssh_key_or_prompt(spark_client.secrets_configuration.ssh_pub_key,
-                                                        user_configuration.username, user_configuration.password,
-                                                        spark_client.secrets_configuration)
+        ssh_key, password = utils.get_ssh_key_or_prompt(
+            spark_client.secrets_configuration.ssh_pub_key,
+            user_configuration.username,
+            user_configuration.password,
+            spark_client.secrets_configuration,
+        )
         cluster_conf.user_configuration = aztk.spark.models.UserConfiguration(
             username=user_configuration.username, password=password, ssh_key=ssh_key)
     else: