+++ aztk_cli/spark/endpoints/cluster/cluster_get.py
@@ -6,11 +6,14 @@ from aztk_cli import config, log, utils
 
 
 def setup_parser(parser: argparse.ArgumentParser):
-    parser.add_argument('--id', dest='cluster_id', required=True, help='The unique id of your spark cluster')
-    parser.add_argument('--show-config', dest='show_config', action='store_true', help='Show the cluster configuration')
-    parser.add_argument('--internal', action='store_true',
-                        help="Show the local IP of the nodes. "\
-                             "Only use if using connecting with a VPN.")
+    parser.add_argument("--id", dest="cluster_id", required=True, help="The unique id of your spark cluster")
+    parser.add_argument("--show-config", dest="show_config", action="store_true", help="Show the cluster configuration")
+    parser.add_argument(
+        "--internal",
+        action="store_true",
+        help="Show the local IP of the nodes. "
+        "Only use if using connecting with a VPN.",
+    )
     parser.set_defaults(internal=False)
 
 
@@ -20,8 +23,9 @@ def execute(args: typing.NamedTuple):
     cluster = spark_client.cluster.get(cluster_id)
     utils.print_cluster(spark_client, cluster, args.internal)
 
-    configuration = spark_client.cluster.get_cluster_config(cluster_id)
-    if configuration and args.show_config:
-        log.info("-------------------------------------------")
-        log.info("Cluster configuration:")
-        utils.print_cluster_conf(configuration, False)
+    if args.show_config:
+        configuration = spark_client.cluster.get_configuration(cluster_id)
+        if configuration:
+            log.info("-------------------------------------------")
+            log.info("Cluster configuration:")
+            utils.print_cluster_conf(configuration, False)